Refactor: Extract shared benchmark utilities and fix CI test paths
- Create benchmarks/utils.py with 10 shared functions (load_mnist_data,
  train_step, evaluate, benchmark_model, plot_comparison, etc.)
- Update bench_mnist_hyperboloid.py and bench_mnist_poincare.py to
  import from utils instead of duplicating ~200 lines each
- Add c parameter to benchmark_model() to handle curvature differences
- Fix .github/workflows/ci.yaml to reference tests/nn_layers/ subdirectory
  for NN layer tests (matching commit 225cbe7 refactor)
- Add nn_layers/test_poincare_conv.py to CI matrix (was missing)
