+++ .github/workflows/ci.yaml
@@ -76,15 +76,18 @@ jobs:
       fail-fast: false
       matrix:
         test-suite:
+          # Top-level tests
           - "test_helpers.py"
-          - "test_hyperboloid_activations.py"
-          - "test_hyperboloid_conv.py"
           - "test_manifolds.py"
           - "test_math_utils.py"
-          - "test_nn_layers.py"
           - "test_optimizers.py"
-          - "test_regression_layers.py"
           - "test_wrapped_normal.py"
+          # NN layer tests (moved to subdirectory)
+          - "nn_layers/test_hyperboloid_activations.py"
+          - "nn_layers/test_hyperboloid_conv.py"
+          - "nn_layers/test_nn_layers.py"
+          - "nn_layers/test_regression_layers.py"
+          - "nn_layers/test_poincare_conv.py"
 
     steps:
       - uses: actions/checkout@v4
